Reorganize `onstart` test programs as requested
