+++ test/program-onstart.test.js
@@ -7,7 +7,7 @@ const runProgram = require('./run-program');
 describe('Test the onStart option', () => {
 	test('Works when given a function', async () => {
 		await runProgram(
-			`${__dirname}/programs/onstart/good/cli/entry.js`,
+			`${__dirname}/programs/onstart-good/cli/entry.js`,
 			'list toppings --sort alphabetical'
 		).then(resolved => {
 			expect(resolved.stdout).toContain('This is the onStart function');
@@ -16,7 +16,7 @@ describe('Test the onStart option', () => {
 
 	test('Fails when given a string', async () => {
 		await runProgram(
-			`${__dirname}/programs/onstart/bad/cli/entry.js`,
+			`${__dirname}/programs/onstart-bad/cli/entry.js`,
 			'list toppings --sort alphabetical'
 		).then(
 			() => {
