+++ test/programs/onstart-bad/cli/entry.js
@@ -1,5 +1,5 @@
 // Import and initialize Waterfall CLI
-const waterfall = require('../../../../../dist/index');
+const waterfall = require('../../../../dist/index');
 
 const onStart = 'This is not a function';
