+++ test/programs/onstart-bad/cli/entry.json
