+++ test/programs/onstart-bad/cli/list/list.json
