+++ test/programs/onstart-bad/package.json
