+++ test/programs/onstart-good/cli/entry.js
@@ -1,5 +1,5 @@
 // Import and initialize Waterfall CLI
-const waterfall = require('../../../../../dist/index');
+const waterfall = require('../../../../dist/index');
 
 const onStart = function() {
 	console.log('This is the onStart function');
