+++ test/programs/onstart-good/cli/entry.json
