+++ test/programs/onstart-good/cli/list/list.json
