+++ test/programs/onstart-good/package.json
